+++ playground/parrots/CMakeLists.txt
@@ -6,7 +6,6 @@
 set(SOURCES
     Overlay.h Overlay.cpp
     BlurringNode.h BlurringNode.cpp
-    SceneTexture.h SceneTexture.cpp
     main.cpp)
 
 qt_add_resources(SOURCES images.qrc shaders.qrc)
